requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
Pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
